finish AuthModule

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { TasksRepository } from './tasks.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,32 +15,32 @@ export class TasksService {
         private tasksRepository: TasksRepository) { }
 
 
-    getTasks(filterDto: GetTaskFilterDto) {
-        return this.tasksRepository.getTasks(filterDto);
+    getTasks(filterDto: GetTaskFilterDto, user: User) {
+        return this.tasksRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.tasksRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.tasksRepository.findOne({ where: { id, userId: user.id } });
         if (!found) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.tasksRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.tasksRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.tasksRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.tasksRepository.delete({ id, userId: user.id });
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
     }
 
-    async updateTask(id: number, taskDto: TaskDto): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTask(id: number, taskDto: TaskDto, user): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         const updated = {
             ...task,
             ...taskDto
@@ -47,9 +48,10 @@ export class TasksService {
         await this.tasksRepository.save(updated);
         return updated;
 
+
     }
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await this.tasksRepository.save(task);
         return task;
